requirements.txt
python-dotenv==1.0.0
pillow==12.0.0
instabot==0.117.0
